rename package to client, add authnz checks
