+++ pkg/client/types.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package openstack
+package client
 
 import (
 	"github.com/gophercloud/gophercloud"
